pyproject.toml
[project]
name = "primeiro-trabalho"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "networkx>=3.5",
    "openai>=2.6.0",
    "pathlib>=1.0.1",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "pyvis>=0.3.2",
    "streamlit>=1.50.0",
    "wordcloud>=1.9.4",
]
